requirements.txt
Django==1.10.5
djangorestframework==3.5.3
httpie==0.9.9
Pygments==2.1.3
requests==2.12.4
